fix: 类型匹配上的错误

+++ utility/Condition.cpp
@@ -13,7 +13,7 @@ bool zhengqi::utility::Condition::waitForSeconds(double seconds)
     clock_gettime(CLOCK_REALTIME, &abstime);
 
     const int64_t kNanoSecondsPerSecond = 1000000000;
-    int nanoseconds = static_cast<int64_t>(seconds * kNanoSecondsPerSecond);
+    int64_t nanoseconds = static_cast<int64_t>(seconds * kNanoSecondsPerSecond);
 
     abstime.tv_sec += static_cast<time_t>((abstime.tv_nsec + nanoseconds) / kNanoSecondsPerSecond);
     abstime.tv_nsec = static_cast<long>((abstime.tv_nsec + nanoseconds) % nanoseconds);
